Bing proxy to controller, data directory storage.
Moved Bing proxy/metadata to controller and store the
cached data in the data directory.
